requirements.txt
streamlit
pandas
numpy
scikit-learn
PyJWT
bcrypt
mysql-connector-python
plotly